+++ instrumentation/opentelemetry-instrumentation-urllib3/src/opentelemetry/instrumentation/urllib3/__init__.py
@@ -90,7 +90,15 @@ _SUPPRESS_HTTP_INSTRUMENTATION_KEY = context.create_key(
 
 _UrlFilterT = typing.Optional[typing.Callable[[str], str]]
 _RequestHookT = typing.Optional[
-    typing.Callable[[Span, urllib3.connectionpool.HTTPConnectionPool], None]
+    typing.Callable[
+        [
+            Span,
+            urllib3.connectionpool.HTTPConnectionPool,
+            typing.Dict,
+            typing.Optional[str],
+        ],
+        None,
+    ]
 ]
 _ResponseHookT = typing.Optional[
     typing.Callable[
@@ -150,6 +158,7 @@ def _instrument(
         method = _get_url_open_arg("method", args, kwargs).upper()
         url = _get_url(instance, args, kwargs, url_filter)
         headers = _prepare_headers(kwargs)
+        body = _get_url_open_arg("body", args, kwargs)
 
         span_name = "HTTP {}".format(method.strip())
         span_attributes = {
@@ -161,7 +170,7 @@ def _instrument(
             span_name, kind=SpanKind.CLIENT, attributes=span_attributes
         ) as span:
             if callable(request_hook):
-                request_hook(span, instance)
+                request_hook(span, instance, headers, body)
             inject(headers)
 
             with _suppress_further_instrumentation():